finance_dashboard(1)/finance_dashboard/finance_dashboard/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
marshmallow==3.21.3
python-dotenv==1.0.1
